Change initramfs path to relative path to an symbol link.
If we use relative path for initramfs in .config, we have to
modify the .config file everytime we want to buile image from
recovery to android boot.

Use the symbol link, it would be easier by just change the link
to a new location.

ln -s ../out/target/product/hd806/recovery/root root